Fix: Double precision operation in render kernel
Data type for variable ir in delectric class was declared as double.
Replaced this to be a single precision float variable as double
precision operations running on a GPU can significantly impact
performance.

+++ material.h
@@ -114,7 +114,7 @@ public:
 		return true;
 	}
 
-	double ir;
+	float ir;
 
 private:
 	__device__ static float reflectance(float cosine, float ref_index)
